requirements.txt
absl-py==1.4.0
alembic==1.9.2
attrs==22.2.0
autopep8==2.0.1
boto3==1.26.54
botocore==1.29.54
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-S3==0.3.3
Flask-SQLAlchemy==3.0.2
flatbuffers==23.1.21
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
greenlet==2.0.1
gspread==5.7.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.6.3
mediapipe==0.9.0.1
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.46
SQLAlchemy-serializer==1.4.1
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
